Text-Rendering hinzugefügt
rendert Text als Bitmap solid und die Objekte weiterhin im Wireframe

+++ Kollisionserkennung DirectX/Lightshaderclass.cpp
@@ -301,7 +301,7 @@ void LightShaderClass::OutputShaderErrorMessage(ID3D10Blob* errorMessage, HWND h
 	compileErrors = (char*)(errorMessage->GetBufferPointer());
 
 	// Get the length of the message.
-	bufferSize = errorMessage->GetBufferSize();
+	bufferSize = (unsigned long)errorMessage->GetBufferSize();
 
 	// Open a file to write the error message to.
 	fout.open("shader-error.txt");
